Better separation of plugin implementations.
