+++ crates/dprint/src/plugins/implementations/mod.rs
@@ -1,0 +1,7 @@
+mod process;
+mod wasm;
+mod common;
+
+pub use common::*;
+
+pub use wasm::compile as compile_wasm;
