+++ crates/dprint/src/plugins/implementations/wasm/load_instance.rs
