+++ crates/dprint/src/plugins/mod.rs
@@ -1,9 +1,7 @@
-pub mod wasm;
-mod process;
+mod implementations;
 mod plugin;
 mod cache;
 mod cache_manifest;
-mod common;
 mod resolver;
 mod pool;
 mod repo;
@@ -11,9 +9,10 @@ mod types;
 
 pub use plugin::*;
 pub use cache::*;
-use common::*;
 use cache_manifest::*;
 pub use resolver::*;
 pub use pool::*;
 pub use repo::*;
 pub use types::*;
+
+pub use implementations::compile_wasm;
